gateway/requirements.txt
fastapi
uvicorn
websockets
grpcio
redis
protobuf